Refactor File interface
