+++ nexus-writer/src/nexus/file_interface/mock_nexus_file.rs
@@ -1,13 +1,10 @@
 use std::path::Path;
 
 use crate::{
-    hdf5_handlers::NexusHDF5Result,
-    nexus_structure::Root,
-    run_engine::{run_messages::HandlesAllNexusMessages, RunParameters},
-    NexusSettings,
+    hdf5_handlers::NexusHDF5Result, nexus::NexusMessageHandler, nexus_structure::Root, run_engine::{run_messages::HandlesAllNexusMessages, RunParameters}, NexusSettings
 };
 
-use super::{NexusFileInterface, NexusMessageHandler};
+use super::NexusFileInterface;
 
 pub(crate) struct NexusNoFile;
